refactor: 가격 관련 변수들을 Long으로 변경

+++ app/src/main/java/woowacourse/shopping/data/model/response/product/ProductContent.kt
@@ -8,6 +8,6 @@ data class ProductContent(
     @SerialName("id") val id: Long,
     @SerialName("name") val name: String,
     @SerialName("imageUrl") val imageUrl: String,
-    @SerialName("price") val price: Int,
+    @SerialName("price") val price: Long,
     @SerialName("category") val category: String,
 )
