+++ app/src/main/java/woowacourse/shopping/domain/CartItem.kt
@@ -5,6 +5,6 @@ data class CartItem(
     val product: Product,
     val quantity: Int,
 ) {
-    val totalPrice: Int
+    val totalPrice: Long
         get() = product.price.value * quantity
 }
