+++ app/src/main/java/woowacourse/shopping/domain/coupon/PercentageCoupon.kt
@@ -17,8 +17,8 @@ data class PercentageCoupon(override val info: CouponInfo, override val type: Co
         return true
     }
 
-    override fun discount(cartItems: List<CartItem>): Int {
+    override fun discount(cartItems: List<CartItem>): Long {
         val discountRate = (info.discount ?: 0) / 100.0
-        return (cartItems.sumOf(CartItem::totalPrice) * discountRate).toInt()
+        return (cartItems.sumOf(CartItem::totalPrice) * discountRate).toLong()
     }
 }
